requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
Authlib==1.3.0
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
